Add `.eslintrc.json` ESLint config file
Next to using the ESLint recommended settings, the configuration is using WordPress' ESLint plugin to adhere to WordPress Coding Standards.
During the pre-stable phase of this plugin, console logging will be ignored.

Note: this config file is using a syntax compatible with ESLint up to v8. It is **not** fully tested or known to be compatible with ESLint 9.x.