Add `last-updated` data attribute to both result item elements
The attribute is populated with the local timestamp in ISO 8601 format, based on the current time zone configured in the WordPress website.

The ISO 8601 date timestamp can now be retrieved for all items, using the `Item::get_last_updated()` method. Note that by default the function will use the timezone provided by the RSD API.